requirements.txt
azure-functions
azure-eventhub
python-dateutil
requests